+++ reed_solomon.go
@@ -145,8 +145,8 @@ func (rss *reedSolomonSplitter) NextBytes() ([]byte, error) {
 }
 
 // Chunksize returns the chunk size of this Splitter.
-func (rss *reedSolomonSplitter) Chunksize() uint32 {
-	return uint32(rss.size)
+func (rss *reedSolomonSplitter) Chunksize() uint64 {
+	return uint64(rss.size)
 }
 
 // Splitters returns the underlying individual splitters.
